Split out module-lattice module into individual files
